delete old feed and change query for feed

+++ tkpdfeed/src/main/java/com/tokopedia/tkpd/tkpdfeed/feedplus/domain/model/feeddetail/FeedDetailProductDomain.java
@@ -55,7 +55,7 @@ public class FeedDetailProductDomain {
 
     private final
     @Nullable
-    Integer rating;
+    Double rating;
 
     public FeedDetailProductDomain(Integer id,
                                    String name,
@@ -68,7 +68,7 @@ public class FeedDetailProductDomain {
                                    Object url,
                                    String productLink,
                                    Boolean wishlist,
-                                   Integer rating) {
+                                   Double rating) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -139,7 +139,7 @@ public class FeedDetailProductDomain {
     }
 
     @Nullable
-    public Integer getRating() {
+    public Double getRating() {
         return rating;
     }
 }
